Adding final project
I have added the final project.
Project description: The objective of the project is to use web tecnologies such as JQuery, Ajax, Javascript, PHP, HTML5 and CSS to allow new students to register in a database. The system also provides a CRUD session for an administrator which facilitates the registers of the new students in the database. The system also generates pdf to the user after the registration.
All the html, php and css files for the project were addes, as well as the database script